+++ Backend/group-service/internal/app/models/group_invitation.go
@@ -9,9 +9,9 @@ import (
 // GroupInvitation представляет модель приглашения в группу в базе данных
 type GroupInvitation struct {
 	gorm.Model
-	GroupID         int64     `gorm:"not null;index"`  // ID группы, куда приглашают
+	GroupID         uint      `gorm:"not null;index"`  // ID группы, куда приглашают
 	InvitationCode  string    `gorm:"unique;not null"` // Уникальный код приглашения
-	CreatedByUserID int64     `gorm:"not null"`        // ID пользователя, создавшего приглашение (админ группы)
+	CreatedByUserID uint      `gorm:"not null"`        // ID пользователя, создавшего приглашение (админ группы)
 	ExpiresAt       time.Time `gorm:"not null"`        // Время истечения срока действия приглашения
 
 	// Связь с группой
